ImprovePerformance

+++ Backend/src/main/java/com/improvement_app/food/domain/dietsummary/MealIngredientDTO.java
@@ -1,4 +1,4 @@
 package com.improvement_app.food.domain.dietsummary;
 
-public record MealIngredientDTO(String name, String unit, int amount, int id, int productId) {
+public record MealIngredientDTO(String name, String unit, double amount, long id, long productId) {
 }
